+++ packages/core/src/HeaderRow.js
@@ -1,7 +1,6 @@
-// @flow
 import React from 'react';
+import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import type { ReactChildren } from 'react-flow-types';
 import './styles/HeaderRow.scss';
 
 import prefixClass from './utils/prefixClass';
@@ -16,13 +15,31 @@ export const BEM = {
     right: ROOT_BEM.element('right'),
 };
 
-export type Props = {
-    left?: ReactChildren,
-    center?: ReactChildren,
-    right?: ReactChildren,
-    className?: string, // eslint-disable-line react/require-default-props
-    children?: ReactChildren,
+// --------------------
+//  Helper Component
+// --------------------
+
+export function HeaderArea({ content, ...props }) {
+    if (content === false) {
+        return null;
+    }
+    return <div {...props}>{content}</div>;
+}
+
+HeaderArea.propTypes = {
+    content: PropTypes.oneOfType([
+        PropTypes.node,
+        PropTypes.oneOf([false]),
+    ]),
 };
+
+HeaderArea.defaultProps = {
+    content: undefined,
+};
+
+// --------------------
+//  Main Component
+// --------------------
 
 function HeaderRow({
     left,
@@ -32,7 +49,7 @@ function HeaderRow({
     className,
     children,
     ...otherProps,
-}: Props) {
+}) {
     const rootClassName = classNames(
         BEM.root.toString(),
         className,
@@ -40,13 +57,19 @@ function HeaderRow({
 
     return (
         <div className={rootClassName} {...otherProps}>
-            <div className={BEM.left}>{left}</div>
-            <div className={BEM.center}>{center}</div>
-            <div className={BEM.right}>{right}</div>
+            <HeaderArea content={left} className={BEM.left} />
+            <HeaderArea content={center} className={BEM.center} />
+            <HeaderArea content={right} className={BEM.right} />
             {children}
         </div>
     );
 }
+
+HeaderRow.propTypes = {
+    left: HeaderArea.propTypes.content,
+    center: HeaderArea.propTypes.content,
+    right: HeaderArea.propTypes.content,
+};
 
 HeaderRow.defaultProps = {
     left: undefined,