+++ packages/core/src/__tests__/List.test.js
@@ -1,8 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { shallow } from 'enzyme';
+import { shallow, mount } from 'enzyme';
 
-import List, { BEM as LIST_BEM } from '../List';
+import List from '../List';
+import Section from '../Section';
+
+import ListSpacingContext from '../contexts/listSpacing';
 
 it('renders without crashing', () => {
     const div = document.createElement('div');
@@ -11,38 +14,45 @@ it('renders without crashing', () => {
     ReactDOM.render(element, div);
 });
 
-it('renders in variants with "normal" as default', () => {
+it('consumes context to render a <Section> with spacing configs', () => {
     const wrapper = shallow(<List>Foo Bar</List>);
-    expect(wrapper.hasClass(`${LIST_BEM.root.modifier('normal')}`)).toBeTruthy();
+    expect(wrapper.is(ListSpacingContext.Consumer)).toBeTruthy();
+
+    const renderedElement = wrapper.prop('children')(true);
+    expect(renderedElement.type).toBe(Section);
+    expect(renderedElement.props.verticalSpacing).toBe(true);
+    expect(renderedElement.props.bodySpacing).toBe(false);
+});
+
+it('renders a <ul> inside root <Section>', () => {
+    const wrapper = mount(<List>Foo Bar</List>);
+
+    expect(wrapper.find(Section).find('ul').exists()).toBeTruthy();
+    expect(wrapper.find(Section).find('ul').text()).toBe('Foo Bar');
+});
+
+it('renders in variants with "normal" as default', () => {
+    const wrapper = mount(<List>Foo Bar</List>);
+    expect(wrapper.find(Section).hasClass('gyp-list--normal')).toBeTruthy();
 
     wrapper.setProps({ variant: 'setting' });
-    expect(wrapper.hasClass(`${LIST_BEM.root.modifier('setting')}`)).toBeTruthy();
+    expect(wrapper.find(Section).hasClass('gyp-list--setting')).toBeTruthy();
 
     wrapper.setProps({ variant: 'button' });
-    expect(wrapper.hasClass(`${LIST_BEM.root.modifier('button')}`)).toBeTruthy();
+    expect(wrapper.find(Section).hasClass('gyp-list--button')).toBeTruthy();
 });
 
-it('renders children inside a <ul>', () => {
-    const wrapper = shallow(<List>Foo Bar</List>);
+it('passes unknown props to wrapper <Section>', () => {
+    const wrapper = mount(
+        <List
+            id="foo"
+            verticalSpacing={false} />
+    );
 
-    expect(wrapper.find('ul')).toHaveLength(1);
-    expect(wrapper.find('ul').text()).toBe('Foo Bar');
-});
-
-it('renders title in <div> when specified', () => {
-    const wrapper = shallow(<List>Foo</List>);
-    expect(wrapper.find(`.${LIST_BEM.title}`).exists()).toBeFalsy();
-
-    wrapper.setProps({ title: 'Bar' });
-    expect(wrapper.find(`.${LIST_BEM.title}`)).toHaveLength(1);
-    expect(wrapper.find(`.${LIST_BEM.title}`).text()).toBe('Bar');
-});
-
-it('renders desc in <div> when specified', () => {
-    const wrapper = shallow(<List>Foo</List>);
-    expect(wrapper.find(`.${LIST_BEM.desc}`).exists()).toBeFalsy();
-
-    wrapper.setProps({ desc: <span data-test="foo" /> });
-    expect(wrapper.find(`.${LIST_BEM.desc}`)).toHaveLength(1);
-    expect(wrapper.find(`.${LIST_BEM.desc}`).find('[data-test="foo"]')).toHaveLength(1);
+    expect(wrapper.find(Section).props()).toEqual(
+        expect.objectContaining({
+            id: 'foo',
+            verticalSpacing: false,
+        })
+    );
 });