+++ packages/core/src/__tests__/Popover.test.js
@@ -3,6 +3,7 @@ import ReactDOM from 'react-dom';
 import { shallow } from 'enzyme';
 
 import Popover, { PurePopover, BEM } from '../Popover';
+import ListSpacingContext from '../contexts/listSpacing';
 
 describe('<Popover> with mixins', () => {
     it('should render without crashing', () => {
@@ -14,16 +15,23 @@ describe('<Popover> with mixins', () => {
 });
 
 describe('Pure <Popover>', () => {
+    it('provides context for children list spacing', () => {
+        const wrapper = shallow(<PurePopover />);
+
+        expect(wrapper.find(ListSpacingContext.Provider).exists()).toBeTruthy();
+        expect(wrapper.find(ListSpacingContext.Provider).prop('value')).toBe(false);
+    });
+
     it('renders class names in respond to placement prop', () => {
         const wrapper = shallow(<PurePopover placement="bottom" />);
-        expect(wrapper.hasClass(
+        expect(wrapper.children('div').hasClass(
             BEM.root
                 .modifier('bottom')
                 .toString({ stripBlock: true })
         )).toBeTruthy();
 
         wrapper.setProps({ placement: 'top' });
-        expect(wrapper.hasClass(
+        expect(wrapper.children('div').hasClass(
             BEM.root
                 .modifier('top')
                 .toString({ stripBlock: true })