+++ packages/core/src/contexts/listSpacing.js
@@ -1,0 +1,6 @@
+// #FIXME: use React navive context API when upgraded to v16+
+import createReactContext from 'create-react-context';
+
+const ListSpacingContext = createReactContext(true);
+
+export default ListSpacingContext;