+++ packages/storybook/examples/Form.SwitchRow/index.js
@@ -1,13 +1,12 @@
-import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withInfo } from '@storybook/addon-info';
 
-// For props table
 import SwitchRow from '@ichef/gypcrete-form/src/SwitchRow';
 
 import BasicUsage from './BasicUsage';
+import getPropTables from '../../utils/getPropTables';
 
 storiesOf('[Form] SwitchRow', module)
     .add('basic usage', withInfo()(BasicUsage))
     // Props table
-    .addPropsTable(() => <SwitchRow />);
+    .add('props', getPropTables([SwitchRow]));