+++ packages/storybook/examples/HeaderRow/index.js
@@ -1,15 +1,20 @@
-import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withInfo } from '@storybook/addon-info';
 
-// For props table
 import HeaderRow from '@ichef/gypcrete/src/HeaderRow';
 
 import BasicUsage from './BasicUsage';
+import OptionalArea from './OptionalArea';
+
+import getPropTables from '../../utils/getPropTables';
+
 
 storiesOf('HeaderRow', module)
-    .add('Basic usage',
-        withInfo()(BasicUsage)
+    .add('Basic usage', withInfo()(BasicUsage))
+    .add(
+        'Optional area',
+        withInfo('Remove an area from DOM by explictly setting it to false.')(OptionalArea)
     )
+
     // Props table
-    .addPropsTable(() => <HeaderRow />);
+    .add('props', getPropTables([HeaderRow]));