+++ packages/storybook/examples/TextInput/index.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withInfo } from '@storybook/addon-info';
 
@@ -8,6 +7,8 @@ import TextInput, { PureTextInput } from '@ichef/gypcrete/src/TextInput';
 import BasicUsage from './BasicUsage';
 import WithStatus from './WithStatus';
 import MultiLines from './MultiLines';
+
+import getPropTables from '../../utils/getPropTables';
 
 storiesOf('TextInput', module)
     .add('Basic usage',
@@ -20,4 +21,4 @@ storiesOf('TextInput', module)
         withInfo()(MultiLines)
     )
     // Props table
-    .addPropsTable(() => <TextInput />, [PureTextInput]);
+    .add('props', getPropTables([PureTextInput, TextInput]));